+++ lib/postzord/dispatcher.rb
@@ -4,6 +4,8 @@
 
 
 class Postzord::Dispatcher
+  include Diaspora::Logging
+
   require 'postzord/dispatcher/private'
   require 'postzord/dispatcher/public'
 
@@ -124,7 +126,8 @@ class Postzord::Dispatcher
       batch_deliver_to_local(people)
     else
       people.each do |person|
-        Rails.logger.info("event=push route=local sender=#{@sender.diaspora_handle} recipient=#{person.diaspora_handle} payload_type=#{@object.class}")
+        logger.info "event=push route=local sender=#{@sender.diaspora_handle} recipient=#{person.diaspora_handle} " \
+                    "payload_type=#{@object.class}"
         Workers::Receive.perform_async(person.owner_id, @xml, @sender.person_id)
       end
     end
@@ -134,11 +137,12 @@ class Postzord::Dispatcher
   def batch_deliver_to_local(people)
     ids = people.map{ |p| p.owner_id }
     Workers::ReceiveLocalBatch.perform_async(@object.class.to_s, @object.id, ids)
-    Rails.logger.info("event=push route=local sender=#{@sender.diaspora_handle} recipients=#{ids.join(',')} payload_type=#{@object.class}")
+    logger.info "event=push route=local sender=#{@sender.diaspora_handle} recipients=#{ids.join(',')} " \
+                "payload_type=#{@object.class}"
   end
 
   def deliver_to_hub
-    Rails.logger.debug("event=post_to_service type=pubsub sender_handle=#{@sender.diaspora_handle}")
+    logger.debug "event=post_to_service type=pubsub sender_handle=#{@sender.diaspora_handle}"
     Workers::PublishToHub.perform_async(@sender.public_url)
   end