+++ lib/postzord/receiver/private.rb
@@ -44,7 +44,8 @@ class Postzord::Receiver::Private < Postzord::Receiver
   def receive_object
     obj = @object.receive(@user, @author)
     Notification.notify(@user, obj, @author) if obj.respond_to?(:notification_type)
-    logger.info "user:#{@user.id} successfully received #{@object.class} from person #{@sender.guid}: #{@object.inspect}"
+    logger.info "user:#{@user.id} successfully received #{@object.class} from person #{@sender.guid}: " \
+                "#{@object.inspect}"
     obj
   end